requirements.txt
arrow==0.10.0
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dateutils==0.6.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Login==0.4.0
Flask-Redis==0.3.0
Flask-Script==2.0.6
future==0.16.0
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.14.0
marshmallow-sqlalchemy==0.13.2
passlib==1.7.1
phonenumbers==8.8.6
psycopg2==2.7.3.1
pydash==4.2.1
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.1.15
sqlalchemy-mptt==0.2.2
SQLAlchemy-Utils==0.32.18
sqlservice==0.14.1
urllib3==1.22
Werkzeug==0.12.2
